requirements.txt
# LiveKit Agents Framework
livekit-agents>=0.8.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-cartesia>=0.4.0
livekit-api>=0.6.0

# Anthropic Claude
anthropic>=0.39.0

# Database
supabase>=2.0.0

# HTTP Client
httpx>=0.27.0
aiohttp>=3.9.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2024.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0

# Development
black>=24.0.0
isort>=5.13.0
mypy>=1.8.0
